+++ resources/views/projects/users.blade.php
@@ -18,12 +18,14 @@
                 </div>
             </div>
             <div class="col-sm-auto text-sm-end d-flex align-items-center">
-                <div class="action-btn me-2">
-                    {!! Form::open(['method' => 'DELETE', 'route' => ['projects.user.destroy',  [$project->id,$user->id]]]) !!}
-                    <a href="#" class="mx-3 btn btn-sm  align-items-center bs-pass-para bg-danger" data-bs-toggle="tooltip" title="{{__('Delete')}}"><i class="ti ti-trash text-white"></i></a>
-
-                    {!! Form::close() !!}
-                </div>
+                @if (!empty($user) && $user->type != 'company')
+                        <div class="action-btn me-2">
+                            {!! Form::open(['method' => 'DELETE', 'route' => ['projects.user.destroy',  [$project->id,$user->id]]]) !!}
+                            <a href="#" class="mx-3 btn btn-sm  align-items-center bs-pass-para bg-danger" data-bs-toggle="tooltip" title="{{__('Delete')}}"><i class="ti ti-trash text-white"></i></a>
+                            
+                            {!! Form::close() !!}
+                        </div>
+                @endif
             </div>
         </div>
     </li>
